fix(ui): resolve final text layout findings
