+++ crates/horizon-ui/src/badge.rs
@@ -9,8 +9,6 @@ pub(crate) fn paint_badge_background(
     fill: Color32,
     stroke: Option<(Stroke, StrokeKind)>,
 ) {
-    painter.rect_filled(rect, corner_radius, fill);
-    if let Some((stroke, stroke_kind)) = stroke {
-        painter.rect_stroke(rect, corner_radius, stroke, stroke_kind);
-    }
+    let (stroke, stroke_kind) = stroke.unwrap_or((Stroke::NONE, StrokeKind::Inside));
+    let _ = painter.rect(rect, corner_radius, fill, stroke, stroke_kind);
 }
